chore: address CodeRabbit review on #338 (docs clarity + shared test helper)
- docs(memory-backends): clarify the default no-query recall shows only human
  `remember` notes (insights records are recallable via a query or `--all`);
  reconcile the Graphiti MCP-cleanup note (`forge setup` writes the graphiti-memory
  entry when Graphiti is enabled, so it isn't always user-added).
- test: extract the duplicated makeProjectRoot + rmrfWithRetry helpers from
  recall.test.js and remember.test.js into a shared test/helpers/kernel-project-root.js
  (per-suite factory preserves the configurable tmp prefix + cleanup retries).
